@sarah-ek 2023-07-20
this formula here seems pretty cryptic, is there some reasoning behind it?
```rust
 let n_threads = std::cmp::max(1, std::cmp::min(max_threads, (total_work - threading_threshold + 1) / threading_threshold));
```