@Narsil 2023-07-20
threading_threshold is what you had before to get `num_threads=1 vs num_thread=all`.

`(total_work - threading_threshold + 1) / threading_threshold`
Is simply `ceil(total_work/threading_threshold)` (To get a heuristic on how many threads this looks ok to share.
`min(max_threads, X)` is to not use more threads than requested
`max(1, X)` is to use at least 1.